refactor: modernize project structure
- Restructured project structure
- Add `GetPackages` batch file
- Update namespace
- Add `JsonProperty` decorators
- Add `Name` property to bridge join map
- Add GitHub workflow actions
