+++ .github/ISSUE_TEMPLATE/essentials-plugin-request.md
@@ -1,0 +1,17 @@
+---
+name: Essentials Plugin Request
+about: Request an Essentials Plugin
+title: "[REQUEST] - "
+labels: plugin request
+assignees: ''
+
+---
+
+** Describe the Plugin **
+What do you want the plugin to do? Is it for a specific device? If it's for a specific device, include the manufacturer and model, along with links to API documentation if available.
+
+** Does it need to be available in SIMPL Windows? **
+If it does, describe the functionality that should be exposed to SIMPL Windows via EISC. This may include an example join map, or other way of documenting the layout of the EISC.
+
+** Additional Features or functionality **
+Describe additional features or functionality that this plugin should have.
