+++ .github/workflows/essentialsplugins-betabuilds-caller.yml
@@ -1,0 +1,22 @@
+name: Beta Build
+
+on: 
+  push:
+    branches-ignore: 
+      - "main"
+
+  workflow_dispatch: 
+    inputs: 
+      branch:
+        description: 'Branch to build'
+        required: true 
+        type: string
+
+jobs:
+  call-workflow:
+    uses: PepperDash/workflow-templates/.github/workflows/essentialsplugins-builds.yml@main
+    secrets: inherit
+    with: 
+      branch: ${{ github.ref_name }}
+      default-branch: "main"
+
