+++ packages.config
@@ -1,0 +1,3 @@
+<packages>
+  <package id="PepperDashEssentials" version="1.16.1" targetFramework="net35" allowedVersions="[1.0,2.0)"/>
+</packages>
