--- pepperdash-generic-colorlightZ6-epi/pepperdash-generic-colorlightZ6-epi/ColorlightZ6Properties.cs
@@ -1,11 +1,0 @@
-﻿using PepperDash.Core;
-
-namespace PepperDash.Essentials.Generic.ColorlightZ6
-{
-    public class ColorlightZ6Properties
-    {
-        public ControlPropertiesConfig Control { get; set; }
-
-        public ushort Id { get; set; }
-    }
-}
