--- pepperdash-generic-colorlightZ6-epi/pepperdash-generic-colorlightZ6-epi/Properties/AssemblyInfo.cs
@@ -1,8 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyTitle("pepperdash_generic_colorlightZ6_epi")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("pepperdash_generic_colorlightZ6_epi")]
-[assembly: AssemblyCopyright("Copyright ©  2020")]
-[assembly: AssemblyVersion("1.0.0.*")]
-
