+++ src/ColorlightZ6.sln
@@ -1,0 +1,20 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 10.00
+# Visual Studio 2008
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "epi-colorlight-Z6", "epi-colorlight-Z6.csproj", "{8F8B7548-7AE3-4899-9B3B-5877767F4984}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8F8B7548-7AE3-4899-9B3B-5877767F4984}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8F8B7548-7AE3-4899-9B3B-5877767F4984}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8F8B7548-7AE3-4899-9B3B-5877767F4984}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8F8B7548-7AE3-4899-9B3B-5877767F4984}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
