+++ src/Properties/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("pepperdash_generic_colorlightZ6_epi")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("pepperdash_generic_colorlightZ6_epi")]
+[assembly: AssemblyCopyright("Copyright ©  2020")]
+[assembly: AssemblyVersion("1.0.0.*")]
+
